Issue # 38: Import and export following is now moved to people tab
