+++ src/app/circles/circles.html
@@ -87,10 +87,6 @@
     </div>
   </mat-card> -->
 
-  <p class="circle-actions">
-    <button mat-stroked-button (click)="importFollowList()">Import Following List</button>
-    <button mat-stroked-button (click)="publishFollowList()">Publish Follow List</button>
-  </p>
 
   <small>Circles is how you organize people you follow. Different circles can have different rules applied and circles is an important way to make the experience more enjoyable.</small>
 </div>
