+++ src/app/circles/circles.ts
@@ -3,12 +3,11 @@ import { Router } from '@angular/router';
 import { ApplicationState } from '../services/applicationstate';
 import { Utilities } from '../services/utilities';
 import { DataValidation } from '../services/data-validation';
-import { Circle, NostrEvent, NostrEventDocument, NostrProfileDocument } from '../services/interfaces';
+import { Circle,  NostrProfileDocument } from '../services/interfaces';
 import { ProfileService } from '../services/profile';
 import { CircleService } from '../services/circle';
 import { CircleDialog } from '../shared/create-circle-dialog/create-circle-dialog';
 import { MatDialog } from '@angular/material/dialog';
-import { ImportFollowDialog, ImportFollowDialogData } from './import-follow-dialog/import-follow-dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { AuthenticationService } from '../services/authentication';
 import { copyToClipboard } from '../shared/utilities';
@@ -103,115 +102,13 @@ export class CirclesComponent {
     return pubkeys;
   }
 
-  private getPublicPublicKeys() {
-    console.log(this.items);
-    console.log(this.following);
-
-    const items: string[] = [];
-
-    for (let i = 0; i < this.items.length; i++) {
-      const circle = this.items[i];
-
-      if (circle.public) {
-        const profiles = this.getFollowingInCircle(circle.id);
-        const pubkeys = profiles.map((p) => p.pubkey);
-        items.push(...pubkeys);
-      }
-    }
-
-    return items;
-  }
+ 
 
   getNpub(hex: string) {
     return this.utilities.getNostrIdentifier(hex);
   }
 
-  async publishFollowList() {
-    const publicPublicKeys = this.getPublicPublicKeys();
 
-    await this.dataService.publishContacts(publicPublicKeys);
-
-    this.snackBar.open(`A total of ${publicPublicKeys.length} was added to your public following list`, 'Hide', {
-      duration: 2000,
-      horizontalPosition: 'center',
-      verticalPosition: 'bottom',
-    });
-  }
-
-  async importFollowList() {
-    const dialogRef = this.dialog.open(ImportFollowDialog, {
-      data: { pubkey: this.appState.getPublicKey() },
-      maxWidth: '100vw',
-      panelClass: 'full-width-dialog',
-    });
-
-    dialogRef.afterClosed().subscribe(async (result: ImportFollowDialogData) => {
-      if (!result) {
-        return;
-      }
-
-      this.snackBar.open('Importing followers process has started', 'Hide', {
-        duration: 2000,
-        horizontalPosition: 'center',
-        verticalPosition: 'bottom',
-      });
-
-      let pubkey = this.utilities.ensureHexIdentifier(result.pubkey);
-
-      this.dataService.downloadNewestContactsEvents([pubkey]).subscribe((event) => {
-        const nostrEvent = event as NostrEventDocument;
-        const publicKeys = nostrEvent.tags.map((t) => t[1]);
-
-        for (let i = 0; i < publicKeys.length; i++) {
-          const publicKey = publicKeys[i];
-
-          this.profileService.follow(publicKey);
-
-          // const profile = await this.profile.getProfile(publicKey);
-
-          // // If the user already exists in our database of profiles, make sure we keep their previous circle (if unfollowed before).
-          // if (profile) {
-          //   await this.profile.follow(publicKeys[i], profile.circle);
-          // } else {
-          //   await this.profile.follow(publicKeys[i]);
-          // }
-        }
-
-        // await this.load();
-
-        // this.ngZone.run(() => {
-        //   this.cd.detectChanges();
-        // });
-      });
-
-      // TODO: Add ability to slowly query one after one relay, we don't want to receive multiple
-      // follow lists and having to process everything multiple times. Just query one by one until
-      // we find the list. Until then, we simply grab the first relay only.
-      // this.subscriptions.push(
-      //   this.feedService.downloadContacts(pubkey).subscribe(async (contacts) => {
-      //     const publicKeys = contacts.tags.map((t) => t[1]);
-
-      //     for (let i = 0; i < publicKeys.length; i++) {
-      //       const publicKey = publicKeys[i];
-      //       const profile = await this.profile.getProfile(publicKey);
-
-      //       // If the user already exists in our database of profiles, make sure we keep their previous circle (if unfollowed before).
-      //       if (profile) {
-      //         await this.profile.follow(publicKeys[i], profile.circle);
-      //       } else {
-      //         await this.profile.follow(publicKeys[i]);
-      //       }
-      //     }
-
-      //     await this.load();
-
-      //     this.ngZone.run(() => {
-      //       this.cd.detectChanges();
-      //     });
-      //   })
-      // );
-    });
-  }
 
   createCircle(): void {
     const dialogRef = this.dialog.open(CircleDialog, {
