+++ src/app/people/people.html
@@ -38,6 +38,10 @@
       </mat-expansion-panel>
     </mat-accordion>
   </div>
+  <p class="people-actions">
+    <button mat-stroked-button (click)="importFollowList()">Import Following List</button>
+    <button mat-stroked-button (click)="publishFollowList()">Publish Follow List</button>
+  </p>
 
   <div class="people-list" *ngIf="sortedItems$ | withStatus | async as items">
     <ng-template [ngIf]="items.value">
@@ -73,4 +77,7 @@
   <!-- </div> -->
 
   <p class="dimmed">People are those you are following. You can use circles to organize your people.</p>
+
+
+
 </div>
