+++ src/app/people/people.ts
@@ -5,7 +5,7 @@ import { Utilities } from '../services/utilities';
 import { relayInit } from 'nostr-tools';
 import * as moment from 'moment';
 import { DataValidation } from '../services/data-validation';
-import { NostrEvent, NostrProfile, NostrProfileDocument } from '../services/interfaces';
+import { Circle, NostrEvent, NostrProfile, NostrEventDocument, NostrProfileDocument } from '../services/interfaces';
 import { ProfileService } from '../services/profile';
 import { map, Observable, Subscription } from 'rxjs';
 import { MatDialog } from '@angular/material/dialog';
@@ -13,6 +13,8 @@ import { CircleDialog } from '../shared/create-circle-dialog/create-circle-dialo
 import { FollowDialog, FollowDialogData } from '../shared/create-follow-dialog/create-follow-dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { NavigationService } from '../services/navigation';
+import { ImportFollowDialog, ImportFollowDialogData } from './import-follow-dialog/import-follow-dialog';
+import { DataService } from '../services/data';
 
 @Component({
   selector: 'app-people',
@@ -27,6 +29,8 @@ export class PeopleComponent {
   showMuted = false;
   showAbout = true;
   showFollowingDate = true;
+  following: NostrProfileDocument[] = [];
+  items: Circle[] = [];
 
   items$ = this.profileService.items$;
 
@@ -62,6 +66,7 @@ export class PeopleComponent {
     public dialog: MatDialog,
     public profileService: ProfileService,
     private validator: DataValidation,
+    private dataService: DataService,
     public utilities: Utilities,
     private router: Router,
     private snackBar: MatSnackBar
@@ -231,6 +236,31 @@ export class PeopleComponent {
   //   }
   // }
 
+  getFollowingInCircle(id?: number) {
+    if (id == null) {
+      return this.following.filter((f) => f.circle == null || f.circle == 0);
+    } else {
+      return this.following.filter((f) => f.circle == id);
+    }
+  }
+  private getPublicPublicKeys() {
+    console.log(this.items);
+    console.log(this.following);
+
+    const items: string[] = [];
+
+    for (let i = 0; i < this.items.length; i++) {
+      const circle = this.items[i];
+
+      if (circle.public) {
+        const profiles = this.getFollowingInCircle(circle.id);
+        const pubkeys = profiles.map((p) => p.pubkey);
+        items.push(...pubkeys);
+      }
+    }
+
+    return items;
+  }
   async addFollow(pubkey: string) {
     if (pubkey.startsWith('nsec')) {
       let sb = this.snackBar.open('This is a private key, not a public key.', 'Hide', {
@@ -269,6 +299,93 @@ export class PeopleComponent {
       setTimeout(() => {
         this.router.navigate(['/p', pubkeys[0]]);
       }, 100);
+    });
+  }
+
+  async publishFollowList() {
+    const publicPublicKeys = this.getPublicPublicKeys();
+
+    await this.dataService.publishContacts(publicPublicKeys);
+
+    this.snackBar.open(`A total of ${publicPublicKeys.length} was added to your public following list`, 'Hide', {
+      duration: 2000,
+      horizontalPosition: 'center',
+      verticalPosition: 'bottom',
+    });
+  }
+
+  async importFollowList() {
+    const dialogRef = this.dialog.open(ImportFollowDialog, {
+      data: { pubkey: this.appState.getPublicKey() },
+      maxWidth: '100vw',
+      panelClass: 'full-width-dialog',
+    });
+
+    dialogRef.afterClosed().subscribe(async (result: ImportFollowDialogData) => {
+      if (!result) {
+        return;
+      }
+
+      this.snackBar.open('Importing followers process has started', 'Hide', {
+        duration: 2000,
+        horizontalPosition: 'center',
+        verticalPosition: 'bottom',
+      });
+
+      let pubkey = this.utilities.ensureHexIdentifier(result.pubkey);
+
+      this.dataService.downloadNewestContactsEvents([pubkey]).subscribe((event) => {
+        const nostrEvent = event as NostrEventDocument;
+        const publicKeys = nostrEvent.tags.map((t) => t[1]);
+
+        for (let i = 0; i < publicKeys.length; i++) {
+          const publicKey = publicKeys[i];
+
+          this.profileService.follow(publicKey);
+
+          // const profile = await this.profile.getProfile(publicKey);
+
+          // // If the user already exists in our database of profiles, make sure we keep their previous circle (if unfollowed before).
+          // if (profile) {
+          //   await this.profile.follow(publicKeys[i], profile.circle);
+          // } else {
+          //   await this.profile.follow(publicKeys[i]);
+          // }
+        }
+
+        // await this.load();
+
+        // this.ngZone.run(() => {
+        //   this.cd.detectChanges();
+        // });
+      });
+
+      // TODO: Add ability to slowly query one after one relay, we don't want to receive multiple
+      // follow lists and having to process everything multiple times. Just query one by one until
+      // we find the list. Until then, we simply grab the first relay only.
+      // this.subscriptions.push(
+      //   this.feedService.downloadContacts(pubkey).subscribe(async (contacts) => {
+      //     const publicKeys = contacts.tags.map((t) => t[1]);
+
+      //     for (let i = 0; i < publicKeys.length; i++) {
+      //       const publicKey = publicKeys[i];
+      //       const profile = await this.profile.getProfile(publicKey);
+
+      //       // If the user already exists in our database of profiles, make sure we keep their previous circle (if unfollowed before).
+      //       if (profile) {
+      //         await this.profile.follow(publicKeys[i], profile.circle);
+      //       } else {
+      //         await this.profile.follow(publicKeys[i]);
+      //       }
+      //     }
+
+      //     await this.load();
+
+      //     this.ngZone.run(() => {
+      //       this.cd.detectChanges();
+      //     });
+      //   })
+      // );
     });
   }
 }
